+++ src/main/java/org/peter_lukas/shirtso/commercial/product/image/ProductToImageController.java
@@ -49,7 +49,7 @@ public class ProductToImageController {
     @PutMapping("/primary/{imageId}")
     @LogExecutionTime
     @RolesAllowed(USER_WRITE)
-    public ResponseEntity<Void> updatePrimaryImage(@PathVariable UUID productId, @PathVariable long imageId) {
+    public ResponseEntity<Void> updatePrimaryImage(@PathVariable UUID productId, @PathVariable int imageId) {
         productImageService.updatePrimaryImageStatus(productId, imageId);
         return ResponseEntity.ok().build();
     }
@@ -57,7 +57,7 @@ public class ProductToImageController {
     @DeleteMapping("/{imageId}")
     @LogExecutionTime
     @RolesAllowed(USER_WRITE)
-    public ResponseEntity<Void> removeImageFromProduct(@PathVariable UUID productId, @PathVariable long imageId) {
+    public ResponseEntity<Void> removeImageFromProduct(@PathVariable UUID productId, @PathVariable int imageId) {
         productImageService.removeImageFromProduct(productId, imageId);
         return ResponseEntity.ok().build();
     }
